move definitions from operator to _operator
